requirements.txt
Dijkstar>=2.5.0
psutil>=5.7.3
wcwidth>=0.2.5
numpy>=1.19.4
Pillow>=8.0.1
pyephem>=3.7.7.0
python_dateutil>=2.8.1
